refactor: moved all container logic to separate module; now expects job string from STDIN/input stream instead of Redis.
