+++ grading-vm/grading_container/build_script/cycle_detector.py
@@ -1,6 +1,6 @@
 from typing import Dict, List, Set
-from grading_job.build_script.json_helpers.grading_script_command import is_bash_command, is_conditional_command
-from validations.grading_job_json_types import GradingScriptCommandJSON
+from grading_container.build_script.json_helpers.grading_script_command import is_conditional_command
+from grading_container.validations.grading_job_json_types import GradingScriptCommandJSON
 
 class CycleDetector:
