+++ grading-vm/grading_container/build_script/preprocess.py
@@ -1,14 +1,13 @@
 from typing import Dict, List
 import os
-from grading_job.build_script.code_file.code_file_info import CodeFileInfo, json_to_code_file_info
-from grading_job.build_script.code_file.processing.code_file_processor import CodeFileProcessor
-from grading_job.build_script.cycle_detector import CycleDetector
-from grading_job.build_script.exceptions import InvalidGradingScriptCommand, NotADAGException
-from grading_job.build_script.json_helpers.grading_script_command import is_bash_command, is_conditional_command
-from grading_job.grading_script.bash_grading_script_command import BashGradingScriptCommand
-from grading_job.grading_script.conditional_grading_script_command import ConditionalGradingScriptCommand, GradingScriptPredicate
-from grading_job.grading_script.grading_script_command import GradingScriptCommand
-from validations.grading_job_json_types import GradingScriptCommandJSON
+from grading_container.build_script.code_file.code_file_info import CodeFileInfo
+from grading_container.build_script.code_file.processing.code_file_processor import CodeFileProcessor
+from grading_container.build_script.exceptions import InvalidGradingScriptCommand, NotADAGException
+from grading_container.grading_script.bash_grading_script_command import BashGradingScriptCommand
+from grading_container.grading_script.conditional_grading_script_command import ConditionalGradingScriptCommand, GradingScriptPredicate
+from grading_container.grading_script.grading_script_command import GradingScriptCommand
+from grading_container.build_script.json_helpers.grading_script_command import is_bash_command, is_conditional_command
+from grading_container.validations.grading_job_json_types import GradingScriptCommandJSON
 
 DEFAULT_COMMAND_TIMEOUT = 60 # 1 minute
