+++ grading-vm/grading_container/validations/schemas/bash_grading_script_command_schema.py
