+++ grading-vm/grading_container/validations/schemas/grading_script_condition_schema.py
