+++ grading-vm/job_retrieval/local/local_grading_job_retriever.py
@@ -1,7 +1,5 @@
-from grading_job.job_retrieval.grading_job_retriever import GradingJobRetriever
-from grading_job.job_retrieval.local.exceptions import LocalGradingJobRetrievalError
-from validations.grading_job_json_types import GradingJobJSON
-import json
+from job_retrieval.grading_job_retriever import GradingJobRetriever
+from job_retrieval.local.exceptions import LocalGradingJobRetrievalError
 
 class LocalGradingJobRetriever(GradingJobRetriever):
 
@@ -9,14 +7,10 @@ class LocalGradingJobRetriever(GradingJobRetriever):
     self.__grading_job_config_path = grading_job_config_path
     super().__init__()
     
-  def retrieve_grading_job(self) -> GradingJobJSON:
+  def retrieve_grading_job(self) -> str:
     try:
       with open(self.__grading_job_config_path, 'r') as json_file:
-        grading_job_json = json.load(json_file)
-        if GradingJobRetriever.is_valid_grading_job(grading_job_json):
-          return grading_job_json
-        else:
-          raise LocalGradingJobRetrievalError("The given JSON in this file was not valid.")
+        return json_file.read()
     except FileNotFoundError:
       raise LocalGradingJobRetrievalError("The given file path did not point to a grading job.")
     except IOError:
